requirements.txt
tensorflow-cpu==2.19.0
numpy==1.26.4
pillow==9.5.0
protobuf==3.20.3
grpcio==1.54.0
gradio==5.33.2
fastapi==0.115.12
uvicorn==0.29.0
